backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
google-generativeai==0.7.1
python-dotenv==1.0.1
sqlalchemy
alembic
openai==1.98.0
httpx==0.27.2
